% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{cv.shim}
\alias{cv.shim}
\title{Cross-validation for shim}
\usage{
cv.shim(x, y, main.effect.names, interaction.names, lambda.beta = NULL,
  lambda.gamma = NULL, nlambda.gamma = 10, nlambda.beta = 10,
  nlambda = 100, threshold = 1e-04, max.iter = 100,
  initialization.type = c("ridge", "univariate"), intercept = TRUE,
  normalize = TRUE, verbose = TRUE, parallel = TRUE,
  type.measure = c("mse"), nfolds = 10)
}
\arguments{
\item{x}{Design matrix of dimension \code{n x q}, where \code{n} is the
number of subjects and q is the total number of variables; each row is an
observation vector. This must include all main effects and interactions as
well, with column names corresponding to the names of the main effects
(e.g. \code{x1, x2, E}) and their interactions (e.g. \code{x1:E, x2:E}).
All columns should be scaled to have mean 0 and variance 1; this is done
internally by the \code{\link{shim}} function.}

\item{y}{response variable (matrix form) of dimension \code{n x 1}}

\item{main.effect.names}{character vector of main effects names. MUST be
ordered in the same way as the column names of \code{x}. e.g. if the column
names of \code{x} are \code{"x1","x2"} then \code{main.effect.names =
c("x1","x2")}}

\item{interaction.names}{character vector of interaction names. MUST be
separated by a colon (e.g. x1:x2), AND MUST be ordered in the same way as
the column names of \code{x}}

\item{lambda.beta}{sequence of tuning parameters for the main effects. If
\code{NULL} (default), this function will automatically calculate a
sequence using the \code{\link{shim_once}} function which will be over a
grid of tuning parameters for gamma as well. If the user specifies a
sequence then this function will not automatically perform the serach over
a grid. You will need to create the grid yourself e.g. repeat the
lambda.gamma for each value of lambda.beta}

\item{lambda.gamma}{sequence of tuning parameters for the interaction
effects. Default is \code{NULL} which means this function will
automatically calculate a sequence of tuning paramters. See
\code{\link{shim_once}} for details on how this sequence is calculated.}

\item{nlambda.gamma}{number of tuning parameters for gamma. This needs to be
specified even for user defined inputs}

\item{nlambda.beta}{number of tuning parameters for beta. This needs to be
specified even for user defined inputs}

\item{nlambda}{total number of tuning parameters. If \code{lambda.beta =
NULL} and \code{lambda.gamma = NULL} then \code{nlambda} should be equal to
\code{nlambda.beta x nlambda.gamma}. This is important to specify
especially when a user defined sequence of tuning parameters is set.}

\item{threshold}{Convergence threshold for coordinate descent. Each
coordinate-descent loop continues until the change in the objective
function after all coefficient updates is less than threshold. Default
value is \code{1e-4}.}

\item{max.iter}{Maximum number of passes over the data for all tuning
parameter values; default is 100.}

\item{initialization.type}{The procedure used to estimate the regression
coefficients and used in the \code{\link{uni_fun}} function. If
\code{"univariate"} then a series of univariate regressions is performed
with the response variable \code{y}. If \code{"ridge"} then ridge
regression is performed using the \code{\link[glmnet]{cv.glmnet}} function
and the tuning parameter is chosen using 10 fold cross validation. The
default is \code{"ridge"}.}

\item{intercept}{Should \code{x} and \code{y} be centered. Default is
\code{TRUE}}

\item{normalize}{Should \code{x} be scaled to have unit variance. Default is
\code{TRUE}}

\item{verbose}{Should iteration number and vector of length \code{nlambda} be
printed to console? Default is \code{TRUE}. 0 represents the algorithm has
not converged for the pair of tuning parameters lambda.beta and
lambda.gamma and 1 means it has converged}

\item{parallel}{If \code{TRUE}, use parallel \code{foreach} to fit each fold.
Must register parallel before hand using the
\code{\link[doMC]{registerDoMC}} function from the \code{doMC} package. See
the example below for details.}

\item{type.measure}{loss to use for cross-validation. Currently only 1
option. The default is \code{type.measure="mse"}, which uses squared-error
for gaussian models}

\item{nfolds}{number of folds - default is 10. Although nfolds can be as
large as the sample size (leave-one-out CV), it is not recommended for
large datasets. Smallest value allowable is \code{nfolds=3}}
}
\description{
Does k-fold cross-validation for shim and determines the optimal
  pair of tuning parameters (\eqn{\lambda_\beta} and \eqn{\lambda_\gamma})
}
\details{
The function runs shim nfolds+1 times; the first to get the tuning
  parameter sequences, and then the remainder to compute the fit with each of
  the folds omitted. The error is accumulated, and the average error and
  standard deviation over the folds is computed. Note also that the results
  of cv.shim are random, since the folds are selected at random using the
  \code{\link{createfolds}} function. Users can reduce this randomness by
  running cv.shim many times, and averaging the error curves.
}
\author{
Sahir Bhatnagar

  Maintainer: Sahir Bhatnagar \email{sahir.bhatnagar@mail.mcgill.ca}
}

